feat: add update endpoint for product and add DeletedImage

+++ src/main/java/com/nashrookie/lavish/dto/request/CreateProductDto.java
@@ -16,6 +16,6 @@ public record CreateProductDto(
                 MultipartFile thumbnailImg,
                 @NotNull Boolean isFeatured,
                 @Min(0) Integer quantity,
-                @NotNull Integer categoryId,
+                @NotNull Long categoryId,
                 List<MultipartFile> images) {
 }
